Creating separate folder and file for Alert component.
