+++ art-elephant/client/src/components/Alert/Alert.js
