+++ art-elephant/client/src/components/MyAccount/MyArtistProfile.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import axios from 'axios';
-import Alert from './Alert';
+import Alert from '../Alert/Alert';
 
 class MyArtistProfile extends Component {
     constructor(props){
